+++ src/pipelines.ts
@@ -24,7 +24,11 @@ export {
 	getJob,
 	getJobNeeds,
 	retryPipeline,
-	cancelPipeline
+	cancelPipeline,
+	retryJob,
+	cancelJob,
+	playJob,
+	commitUrl
 } from './gitlab-api';
 export type { RepoConfig } from './gitlab-api';